+++ ui/src/pages/Profile/AccountInformation.css
@@ -1,3 +1,16 @@
 #avatar {
     height: 100%;
+    width: 100%;
+    background-color: aqua;
+    cursor: pointer;
+}
+
+#initials {
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    font-size: larger;
+    font-weight: bolder;
+    cursor: pointer;
 }
