+++ ui/src/services/authentication.service.js
@@ -10,7 +10,6 @@ const currentUserSubject = new BehaviorSubject(
 export const authenticationService = {
   register,
   login,
-  updateProfile,
   logout,
   currentUser: currentUserSubject.asObservable(),
   get currentUserValue() {
@@ -66,16 +65,6 @@ function login(username, password, rememberMe) {
 
       return user;
     });
-}
-
-function updateProfile(name, username, email) {
-  const requestOptions = {
-    method: 'POST',
-    headers: { 'Content-Type': 'application/json' },
-    body: JSON.stringify({ name, username, email }),
-  };
-
-  return fetch(`/api/updateProfile`, requestOptions).then(handleResponse);
 }
 
 function logout() {
